refactor: shared types
