+++ example/web/src/App.tsx
@@ -37,6 +37,8 @@ const nativeMethod = linkNativeMethod<AppBridge>({
   },
 });
 
+// console.log(window.__bridgeMethods__);
+
 function App() {
   const [message, setMessage] = useState("");
