+++ packages/react-native/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./createWebView";
 export * from "./integrations/bridge";
-export * from "./types/bridge";
 export * from "./types/webview";
 export * from "./useBridge";
+export * from "@webview-bridge/types";
