+++ tsconfig.base.json
@@ -2,7 +2,8 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "@webview-bridge/util": ["shared/util/src"]
+      "@webview-bridge/util": ["shared/util/src"],
+      "@webview-bridge/types": ["shared/util/src/types"]
     },
     "plugins": [
       { "transform": "typescript-transform-paths" },
